Sparse and functional CCA

moma_sfcca creates an SFCCA R6 object and returns it. Type ?CCA_deflation for description of problem formulation and deflation scheme.

moma_scca is a function for performing one-way sparse CCA.

moma_twscca is a function for performing two-way sparse CCA.

moma_fcca is a function for performing one-way functional CCA.

moma_twfcca is a function for performing two-way functional CCA.

moma_sfcca(X, ..., Y, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_scca(X, ..., Y, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_twscca(X, ..., Y, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_fcca(X, ..., Y, center = TRUE, scale = FALSE,
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_twfcca(X, ..., Y, center = TRUE, scale = FALSE,
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

Arguments

X, Y
A data matrix, each row representing a sample, and each column a feature.

...
Force users to specify arguments by names.

center
A logical value indicating whether the variables should be shifted to be zero centered. Defaults to TRUE.

scale
A logical value indicating whether the variables should be scaled to have unit variance. Defaults to FALSE.

x_sparse, y_sparse
An object of class inheriting from "moma_sparsity_type". Most conveniently specified by functions described in moma_sparsity_options. It specifies the type of sparsity-inducing penalty function used in the model. Note that for moma_scca, these two parameters must not be specified at the same time. For moma_fcca and moma_twfcca, they must not be specified.

x_smooth, y_smooth
An object of class inheriting from "moma_smoothness_type". Most conveniently specified by functions described in moma_smoothness. It specifies the type of smoothness terms used in the model. Note that for moma_fcca, these two parameters must not be specified at the same time. For moma_scca and moma_twscca, they must not be specified.

pg_settings
An object of class inheriting from "moma_pg_settings". Most conviently specified by functions described in moma_pg_settings. It specifies the type of algorithm used to solve the problem, acceptable level of precision, and the maximum number of iterations allowed.

max_bic_iter
A positive integer. Defaults to 5. The maximum number of iterations allowed in nested greedy BIC selection scheme.

rank
A positive integer. Defaults to 1. The maximal rank, i.e., maximal number of principal components to be used.

Functions

moma_scca: a function for performing one-way sparse CCA.

moma_twscca: a function for performing two-way sparse CCA

moma_fcca: a function for performing one-way functional CCA

moma_twfcca: a function for performing two-way functional CCA